# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Permy \u2014 Building Permit & Construction Intelligence API",
    "description": "Normalized permits, contractor activity, property timelines, and ZIP development signals \u2014 ranked, sourced, and ready for your app. Give Permy an address, city, ZIP, contractor, or trade and get back clean, machine-readable municipal intelligence with source links and confidence scores.",
    "termsOfService": "https://docs.permy.dev/terms",
    "contact": {
      "name": "Permy",
      "url": "https://permy.dev/",
      "email": "hi@permy.dev"
    },
    "license": {
      "name": "Apache-2.0 (code); data per upstream license",
      "url": "https://docs.permy.dev/legal"
    },
    "version": "0.1.0",
    "x-logo": {
      "url": "https://permy.dev/assets/brand-mark.png"
    }
  },
  "servers": [
    {
      "url": "https://permy.p.rapidapi.com",
      "description": "RapidAPI production"
    },
    {
      "url": "https://api.permy.dev",
      "description": "Direct site"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local dev"
    }
  ],
  "paths": {
...
    "/v1/leads/ranked": {
      "get": {
        "tags": [
          "leads & intelligence"
        ],
        "summary": "Ranked opportunities for a buyer persona with lead_score + recommended_action + reason",
        "operationId": "leads_ranked_v1_leads_ranked_get",
        "parameters": [
          {
            "name": "persona",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(roofer|solar|hvac|investor|supplier|insurer|general)$",
              "default": "general",
              "title": "Persona"
            }
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "zip",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Zip"
            }
          },
          {
            "name": "trade",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trade"
            }
          },
          {
            "name": "min_valuation",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Valuation"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankedLeadsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/intelligence/score": {
      "post": {
        "tags": [
          "leads & intelligence"
        ],
        "summary": "Address or permit_id + persona \u2192 development_score, risk_flags, market narrative, sources",
        "operationId": "intelligence_score_v1_intelligence_score_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntelligenceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntelligenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/v1/coverage": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Supported cities, fields available per city, freshness",
        "operationId": "coverage_v1_coverage_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/v1/usage": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Quota usage for the caller",
        "operationId": "usage_v1_usage_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CoverageResponse": {
        "properties": {
          "cities": {
            "items": {
              "$ref": "#/components/schemas/CoverageCity"
            },
            "type": "array",
            "title": "Cities"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cities",
          "total"
        ],
        "title": "CoverageResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IntelligenceRequest": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "permit_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permit Id"
          },
          "persona": {
            "type": "string",
            "enum": [
              "roofer",
              "solar",
              "hvac",
              "investor",
              "supplier",
              "insurer",
              "general"
            ],
            "title": "Persona",
            "default": "general"
          },
          "project_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "IntelligenceRequest"
      },
      "IntelligenceResponse": {
        "properties": {
          "input": {
            "$ref": "#/components/schemas/IntelligenceRequest"
          },
          "property": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Property"
              },
              {
                "type": "null"
              }
            ]
          },
          "permits": {
            "items": {
              "$ref": "#/components/schemas/Permit"
            },
            "type": "array",
            "title": "Permits"
          },
          "development_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Development Score"
          },
          "permit_activity": {
            "additionalProperties": true,
            "type": "object",
            "title": "Permit Activity"
          },
          "risk_flags": {
            "items": {
              "$ref": "#/components/schemas/RiskFlag"
            },
            "type": "array",
            "title": "Risk Flags"
          },
          "market_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Context"
          },
          "market": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MarketScore"
              },
              {
                "type": "null"
              }
            ]
          },
          "recommended_action": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "call_now",
                  "qualify",
                  "monitor",
                  "skip"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Action"
          },
          "lead_score": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Score"
          },
          "source_links": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Links"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 0.0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "input",
          "development_score"
        ],
        "title": "IntelligenceResponse"
      },
...
      "OwnerRef": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "OwnerRef"
      },
      "Permit": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Permy opaque id (stringified bigint)"
          },
          "canonical_uid": {
            "type": "string",
            "title": "Canonical Uid",
            "description": "Stable hash(jurisdiction + source_permit_id)"
          },
          "jurisdiction_slug": {
            "type": "string",
            "title": "Jurisdiction Slug"
          },
          "source_permit_id": {
            "type": "string",
            "title": "Source Permit Id"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "first_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "First Seen At"
          },
          "last_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen At"
          },
          "last_checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Checked At"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "permit_type_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permit Type Raw"
          },
          "permit_type_normalized": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permit Type Normalized"
          },
          "work_class": {
            "type": "string",
            "enum": [
              "new_construction",
              "alteration",
              "addition",
              "remodel",
              "repair",
              "demolition",
              "other",
              "unknown"
            ],
            "title": "Work Class",
            "default": "unknown"
          },
          "trade_category": {
            "type": "string",
            "enum": [
              "roofing",
              "solar",
              "hvac",
              "plumbing",
              "electrical",
              "building",
              "general",
              "demolition",
              "other",
              "unknown"
            ],
            "title": "Trade Category",
            "default": "unknown"
          },
          "is_new_construction": {
            "type": "boolean",
            "title": "Is New Construction",
            "default": false
          },
          "is_alteration": {
            "type": "boolean",
            "title": "Is Alteration",
            "default": false
          },
          "is_demolition": {
            "type": "boolean",
            "title": "Is Demolition",
            "default": false
          },
          "valuation_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valuation Usd",
            "description": "Declared job value in USD; null when city doesn't publish"
          },
          "housing_units": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Housing Units"
          },
          "new_add_sqft": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Add Sqft"
          },
          "dates": {
            "$ref": "#/components/schemas/PermitDates"
          },
          "current_status": {
            "type": "string",
            "enum": [
              "applied",
              "issued",
              "active",
              "final",
              "expired",
              "cancelled",
              "withdrawn",
              "unknown"
            ],
            "title": "Current Status",
            "default": "unknown"
          },
          "status_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Raw"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "description_enriched": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description Enriched"
          },
          "contractor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContractorRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "owner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OwnerRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "parcel_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parcel Id"
          },
          "enrichment": {
            "$ref": "#/components/schemas/Enrichment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "canonical_uid",
          "jurisdiction_slug",
          "source_permit_id",
          "first_seen_at",
          "last_seen_at",
          "last_checked_at",
          "address"
        ],
        "title": "Permit"
      },
...
      "Property": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "canonical_uid": {
            "type": "string",
            "title": "Canonical Uid"
          },
          "full_address": {
            "type": "string",
            "title": "Full Address"
          },
          "street": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Street"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "zip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zip"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          },
          "geocode_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geocode Confidence"
          },
          "jurisdiction_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction Slug"
          },
          "parcel_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parcel Id"
          },
          "year_built": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year Built"
          },
          "sqft": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sqft"
          },
          "permit_count": {
            "type": "integer",
            "title": "Permit Count",
            "default": 0
          },
          "last_permit_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Permit Date"
          },
          "coverage_status": {
            "type": "string",
            "enum": [
              "covered",
              "partial",
              "no_feed"
            ],
            "title": "Coverage Status",
            "default": "covered"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "canonical_uid",
          "full_address"
        ],
        "title": "Property"
      },
...
      "RankedLead": {
        "properties": {
          "permit": {
            "$ref": "#/components/schemas/Permit"
          },
          "lead_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Lead Score"
          },
          "recommended_action": {
            "type": "string",
            "enum": [
              "call_now",
              "qualify",
              "monitor",
              "skip"
            ],
            "title": "Recommended Action"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "persona": {
            "type": "string",
            "enum": [
              "roofer",
              "solar",
              "hvac",
              "investor",
              "supplier",
              "insurer",
              "general"
            ],
            "title": "Persona"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "permit",
          "lead_score",
          "recommended_action",
          "reason",
          "persona"
        ],
        "title": "RankedLead"
      },
      "RankedLeadsResponse": {
        "properties": {
          "persona": {
            "type": "string",
            "enum": [
              "roofer",
              "solar",
              "hvac",
              "investor",
              "supplier",
              "insurer",
              "general"
            ],
            "title": "Persona"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "leads": {
            "items": {
              "$ref": "#/components/schemas/RankedLead"
            },
            "type": "array",
            "title": "Leads"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "persona",
          "page",
          "limit",
          "total",
          "leads"
        ],
        "title": "RankedLeadsResponse"
      },
      "RiskFlag": {
        "properties": {
          "flag": {
            "type": "string",
            "title": "Flag"
          },
          "severity": {
            "type": "string",
            "enum": [
              "info",
              "warning",
              "critical"
            ],
            "title": "Severity"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "flag",
          "severity"
        ],
        "title": "RiskFlag"
      },
      "UsageResponse": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "day": {
            "type": "string",
            "format": "date",
            "title": "Day"
          },
          "requests_today": {
            "type": "integer",
            "title": "Requests Today"
          },
          "daily_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Limit"
          },
          "monthly_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthly Limit"
          },
          "month_requests": {
            "type": "integer",
            "title": "Month Requests",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "api_key",
          "tier",
          "day",
          "requests_today"
        ],
        "title": "UsageResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RapidAPI-Key",
        "description": "Your RapidAPI app key (sent automatically by the gateway). Direct-site callers may use X-API-Key or Authorization: Bearer instead."
      }
    }
  },
...
}